requirements.txt

pandas
numpy
seaborn
matplotlib
pyarrow
Sqlalchemy
psycopg2-binary
kafka-python
docker